+++ packages/markdown-it-shiki-twoslash/CHANGELOG.md
@@ -1,5 +1,18 @@
 # markdown-it-shiki-twoslash
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [bbba24f]
+  - remark-shiki-twoslash@2.0.5
+
+## 2.0.1
+
+### Patch Changes
+
+- remark-shiki-twoslash@2.0.4
+
 ## 2.0.0
 
 ### Major Changes