+++ packages/twoslash-cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # twoslash-cli
 
+## 1.3.3
+
+### Patch Changes
+
+- Updated dependencies [bbba24f]
+  - remark-shiki-twoslash@2.0.5
+
+## 1.3.2
+
+### Patch Changes
+
+- remark-shiki-twoslash@2.0.4
+
 ## 1.3.1
 
 ### Patch Changes